fix: error handling

+++ controllers/pizza.controller.js
@@ -6,11 +6,23 @@ const pizzaController = {
         const pizzaId = req.params.id;
         const pizza = await pizzaService.getPizzaById(pizzaId);
 
+        if (pizza.length === 0){
+            const err = new Error(`Aucune pizza trouvée`);
+            err.code = 404;
+            throw err;
+        }
+
         const message = `La pizza avec l'id ${pizzaId} a bien été trouvé !`;
         return responseSender.sendSuccessResponse(res, pizza, message);
     },
     getAllPizzas: async (req,res) => {
         const pizzas = await pizzaService.getAllPizzas();
+
+        if (pizzas.length === 0) {
+            const err = new Error(`Aucune pizza trouvée`);
+            err.code = 404;
+            throw err;
+        }
 
         const message = `Toutes les pizzas ont bien été trouvées`;
         return responseSender.sendSuccessResponse(res, pizzas, message);
@@ -20,6 +32,12 @@ const pizzaController = {
 
         const pizza = await pizzaService.getPizzaById(pizzaIdOfTheDay[0].id);
 
+        if (pizza.length === 0){
+            const err = new Error(`Aucune pizza trouvée`);
+            err.code = 404;
+            throw err;
+        }
+
         const message = `La pizza du jour a bien été trouvée.`;
         return responseSender.sendSuccessResponse(res, pizza, message);
     },
@@ -27,6 +45,12 @@ const pizzaController = {
         const pizzaId = req.params.id;
 
         const pizzaAndIngredients = await pizzaService.getIngredientsByPizzaId(pizzaId);
+
+        if (pizzaAndIngredients.length === 0){
+            const err = new Error(`Aucun ingrédients trouvée pour la pizza ${pizzaId}`);
+            err.code = 404;
+            throw err;
+        }
 
         const ingredientIds = [];
 
@@ -49,11 +73,24 @@ const pizzaController = {
     getIngredientById: async (req,res) => {
         const ingredientId = req.params.id;
         const ingredient = await pizzaService.getIngredientById(ingredientId);
+
+        if (ingredient.length === 0){
+            const err = new Error(`Aucun ingrédients trouvé`);
+            err.code = 404;
+            throw err;
+        }
+
         const message = `L'ingrédient avec l'id ${ingredientId} a bien été trouvé !`;
         return responseSender.sendSuccessResponse(res, ingredient, message);
     },
     getAllIngredients: async (req,res) => {
         const ingredients = await pizzaService.getAllIngredients();
+
+        if (ingredients.length === 0){
+            const err = new Error(`Aucun ingrédients trouvé`);
+            err.code = 404;
+            throw err;
+        }
 
         const message = `Tous les ingrédients ont bien été trouvées`;
         return responseSender.sendSuccessResponse(res, ingredients, message);
@@ -62,7 +99,9 @@ const pizzaController = {
         const { name: ingredientName } = req.body;
 
         if (!ingredientName) {
-            return responseSender.sendErrorResponse(res, "Le nom de l'ingrédient est requis.", 400);
+            const err = new Error(`Le nom de l'ingrédient est requis.`);
+            err.code = 400;
+            throw err;
         }
 
         const ingredient = await pizzaService.createNewIngredient(ingredientName);
@@ -74,7 +113,9 @@ const pizzaController = {
         const { name: pizzaName, price, is_offer, ingredients: ingredientNames } = req.body;
 
         if (!pizzaName) {
-            return responseSender.sendErrorResponse(res, "Le nom de la pizza est requis.", 400);
+            const err = new Error(`Le nom de la pizza est requis.`);
+            err.code = 400;
+            throw err;
         }
 
         const newPizza = await pizzaService.createNewPizza(pizzaName, price, is_offer, ingredientNames);
@@ -87,13 +128,19 @@ const pizzaController = {
         const {price: newPrice} = req.body;
 
         if (newPrice === undefined) {
-            return responseSender.sendErrorResponse(res, "Le nouveau prix de la Pizza est requis.", 400);
+            const err = new Error(`Le nouveau prix de la Pizza est requis.`);
+            err.code = 400;
+            throw err;
         }
         if (typeof newPrice !== 'number' || newPrice <= 0) {
-            return responseSender.sendErrorResponse(res, "Le prix doit être un nombre positif.", 400);
+            const err = new Error(`Le prix doit être un nombre positif.`);
+            err.code = 400;
+            throw err;
         }
         if (Math.round(newPrice * 100) % 5 !== 0) {
-            return responseSender.sendErrorResponse(res, "Le prix doit être arrondi aux 5 centimes.", 400);
+            const err = new Error(`Le prix doit être arrondi aux 5 centimes.`);
+            err.code = 400;
+            throw err;
         }
 
         const newPricedPizza = await pizzaService.changePizzaPrice(pizzaId, newPrice);
@@ -106,7 +153,9 @@ const pizzaController = {
         const {name: newName} = req.body;
 
         if (newName === undefined) {
-            return responseSender.sendErrorResponse(res, "Le nouveau nom pour l'ingrédient est requis.", 400);
+            const err = new Error(`Le nouveau nom pour l'ingrédient est requis.`);
+            err.code = 400;
+            throw err;
         }
 
         const newIngredient = await pizzaService.changeIngredientName(ingredientId, newName);
@@ -117,6 +166,7 @@ const pizzaController = {
     deletePizza: async (req,res) => {
         const pizzaId = req.params.id;
         const pizza = await pizzaService.getPizzaById(pizzaId);
+        console.log(`controller: deleting pizza`);
 
         if (!pizza || pizza.length === 0) {
             const err = new Error(`La pizza avec l'ID ${pizzaId} n'existe pas`);
@@ -137,23 +187,23 @@ const pizzaController = {
     deleteIngredient: async (req,res) => {
         const ingredientId = req.params.id;
         const ingredient = await pizzaService.getIngredientById(ingredientId);
+        console.log(`controller: deleting ingredients`);
 
-        if (!ingredientId) {
-            return responseSender.sendErrorResponse(res, "L'ID de l'ingrédient est requis.", 400);
+        if (!ingredient || ingredient.length === 0) {
+            const err = new Error(`L'ingrédient avec l'ID ${ingredientId} n'existe pas`);
+            err.code = 404;
+            throw err;
         }
-        if (ingredient.length <= 0) {
-            return responseSender.sendErrorResponse(res, `L'ingrédient avec l'ID ${ingredientId} n'existe pas`, 404);
-        }
 
-        const deletedPizza = await pizzaService.deleteIngredientById(ingredientId);
+        const deletedIngredient = await pizzaService.deleteIngredientById(ingredientId);
 
-        if (deletedPizza) {
-            const message = `La pizza '${ingredientId}' a bien été supprimée.`;
+        if (deletedIngredient) {
+            const message = `L'ingrédient avec l'ID '${ingredientId}' a bien été supprimée.`;
             return responseSender.sendSuccessResponse(res, ingredientId, message, 200);
-        } else {
-            const message = `Erreur dans la suppression de la pizza ${ingredientId}`;
-            return responseSender.sendErrorResponse(res, message, 400);
         }
+        const err = new Error(`Erreur dans la suppression de l'ingrédient ${ingredientId}`);
+        err.code = 500;
+        throw err;
     }
 };
